Increased i2c settle duration
-also reconciled types for using esp_timer_get_time()

+++ components/dev_ds/celsius_msg.hpp
@@ -36,8 +36,8 @@ public:
     const char *ident;
     Status status = OK;
     float val;
-    int64_t read_us;
-    int64_t convert_us;
+    uint32_t read_us;
+    uint32_t convert_us;
     uint8_t error;
   };
